ProjectをProjectViewに分割
