+++ composables/SFProject.ts
@@ -86,17 +86,3 @@ export class Owner {
   description: string;
   parent: string | null;
 }
-
-export const getOwnerName = (
-  owners: { parent: string; name: string },
-  ownerId: string,
-  offset = 0
-) => {
-  let ret: string[] = [];
-  let current = ownerId;
-  while (current) {
-    ret.unshift(owners[current].name);
-    current = owners[current].parent;
-  }
-  return ret.slice(offset).join(" ");
-};
